add generic
