+++ packages/tldraw/src/lib/tools/SelectTool/childStates/GenericPointingHandle.ts
@@ -1,0 +1,33 @@
+import { StateNode, TLEventHandlers, TLPointerEventInfo, TLShape } from '@tldraw/editor'
+
+export class PointingHandle extends StateNode {
+	info = {} as TLPointerEventInfo & { shape: TLShape; target: 'handle' }
+
+	override onExit = () => {
+		this.editor.setHintingShapes([])
+		this.editor.updateInstanceState(
+			{ cursor: { type: 'default', rotation: 0 } },
+			{ ephemeral: true }
+		)
+	}
+
+	override onPointerUp: TLEventHandlers['onPointerUp'] = () => {
+		this.cancel()
+	}
+
+	override onCancel: TLEventHandlers['onCancel'] = () => {
+		this.cancel()
+	}
+
+	override onComplete: TLEventHandlers['onComplete'] = () => {
+		this.cancel()
+	}
+
+	override onInterrupt = () => {
+		this.cancel()
+	}
+
+	private cancel() {
+		this.parent.transition('idle', this.info)
+	}
+}
