+++ packages/tldraw/src/lib/tools/SelectTool/childStates/PointingLineHandle.ts
@@ -1,9 +1,8 @@
-import { StateNode, TLEventHandlers, TLLineShape, TLPointerEventInfo } from '@tldraw/editor'
+import { TLEventHandlers, TLLineShape, TLPointerEventInfo } from '@tldraw/editor'
+import { PointingHandle } from './GenericPointingHandle'
 
-export class PointingLineHandle extends StateNode {
+export class PointingLineHandle extends PointingHandle {
 	static override id = 'pointing_line_handle'
-
-	info = {} as TLPointerEventInfo & { shape: TLLineShape; target: 'handle' }
 
 	override onEnter = (info: TLPointerEventInfo & { shape: TLLineShape; target: 'handle' }) => {
 		this.info = info
@@ -14,36 +13,9 @@ export class PointingLineHandle extends StateNode {
 		)
 	}
 
-	override onExit = () => {
-		this.editor.updateInstanceState(
-			{ cursor: { type: 'default', rotation: 0 } },
-			{ ephemeral: true }
-		)
-	}
-
-	override onPointerUp: TLEventHandlers['onPointerUp'] = () => {
-		this.parent.transition('idle', this.info)
-	}
-
 	override onPointerMove: TLEventHandlers['onPointerMove'] = () => {
 		if (this.editor.inputs.isDragging) {
 			this.parent.transition('dragging_line_handle', this.info)
 		}
-	}
-
-	override onCancel: TLEventHandlers['onCancel'] = () => {
-		this.cancel()
-	}
-
-	override onComplete: TLEventHandlers['onComplete'] = () => {
-		this.cancel()
-	}
-
-	override onInterrupt = () => {
-		this.cancel()
-	}
-
-	private cancel() {
-		this.parent.transition('idle')
 	}
 }
